Move output code to separate OutputBuffer class
